Refactor CTC alignment into configurable class
