--- scinoephile/audio/transcription/mlx_audio/__init__.py
@@ -1,7 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""MLX-Audio transcription and timestamp alignment.
-
-Package hierarchy (modules may import from any above):
-* forced_alignment
-"""
